Refactor: make components config independents
