+++ robot/simu/SteeringController.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from robot.Component import Component
-from robot.Config import STEERING_COEF
+from robot.simu.Config import STEERING_COEF
 
 
 class SteeringController(Component):
